fixed mongod startup

+++ run.sh
@@ -27,9 +27,13 @@ fi
 ## Setup Mongo and start it
 ##
 mkdir -p $MONGODATABASEDIR	# In case this is the first time ever we are running this 
-rm -f $MONGOOUTFILE
 
-mongod --smallfiles --dbpath $MONGODATABASEDIR --quiet >& $MONGOOUTFILE &
+/etc/init.d/mongod status 2> /dev/null | grep --silent "^mongod (.*) is running..."
+if [ $? -ne 0 ]; then 
+    echo "Starting mongod ..."
+    rm -f $MONGOOUTFILE
+    mongod --smallfiles --dbpath $MONGODATABASEDIR --quiet >& $MONGOOUTFILE &
+fi
 
 
 ##
